renaming samples stimulus controller
