+++ app/controllers/groups/samples_controller.rb
@@ -16,12 +16,7 @@ module Groups
     end
 
     def search
-      request_params = if params[:page].present?
-                         request.request_parameters.slice(:limit, :page)
-                       else
-                         request.request_parameters.slice(:limit)
-                       end
-      redirect_to group_samples_path(request_params)
+      redirect_to group_samples_path(request.request_parameters.slice(:limit, :page))
     end
 
     def select
