+++ app/controllers/projects/samples_controller.rb
@@ -18,12 +18,7 @@ module Projects
     end
 
     def search
-      request_params = if params[:page].present?
-                         request.request_parameters.slice(:limit, :page)
-                       else
-                         request.request_parameters.slice(:limit)
-                       end
-      redirect_to namespace_project_samples_path(request_params)
+      redirect_to namespace_project_samples_path(request.request_parameters.slice(:limit, :page))
     end
 
     def show
