+++ app/views/groups/samples/index.html.erb
@@ -28,7 +28,7 @@
                 },
                 class:
                   "flex items-center px-4 py-2 bg-slate-100 text-slate-600 dark:bg-slate-600 dark:text-slate-300
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                  border-slate-100 dark:border-slate-600 pointer-events-none cursor-not-allowed",
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              border-slate-100 dark:border-slate-600 pointer-events-none cursor-not-allowed",
               ) %>
             <% else %>
               <% dropdown.with_item(
@@ -88,7 +88,7 @@
         <% end %>
       </div>
       <div class="float-right">
-        <%= search_form_for @q, url: search_group_samples_url, html: { method: :post, "data-controller": "filters samples", "data-turbo-action": "replace", "data-samples-page-value": @pagy.page, "data-filters-selection-outlet": "#samples-table", class: "filters" } do |f| %>
+        <%= search_form_for @q, url: search_group_samples_url, html: { method: :post, "data-controller": "filters metadata-toggle", "data-turbo-action": "replace", "data-metadata-toggle-page-value": @pagy.page, "data-filters-selection-outlet": "#samples-table", class: "filters" } do |f| %>
           <div class="flex flex-row-reverse items-center mb-4 space-x-2">
             <input type="hidden" name="format" value="turbo_stream"/>
             <input type="hidden" name="limit" value="<%=@pagy.limit%>"/>
