+++ app/views/projects/samples/index.html.erb
@@ -57,7 +57,7 @@
                 },
                 class:
                   "flex items-center px-4 py-2 bg-slate-100 text-slate-600 dark:bg-slate-600 dark:text-slate-300
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        border-slate-100 dark:border-slate-600 pointer-events-none cursor-not-allowed",
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    border-slate-100 dark:border-slate-600 pointer-events-none cursor-not-allowed",
               ) %>
             <% else %>
               <% dropdown.with_item(
@@ -143,7 +143,7 @@
       <% end %>
     </div>
     <div class="float-right">
-      <%= search_form_for @q, url: search_namespace_project_samples_url, html: { method: :post, "data-controller": "filters samples", "data-turbo-action": "replace", "data-samples-page-value": @pagy.page, "data-filters-selection-outlet": "#samples-table" }, class:
+      <%= search_form_for @q, url: search_namespace_project_samples_url, html: { method: :post, "data-controller": "filters metadata-toggle", "data-turbo-action": "replace", "data-metadata-toggle-page-value": @pagy.page, "data-filters-selection-outlet": "#samples-table" }, class:
         "filters" do |f| %>
         <div class="flex flex-row-reverse items-center mb-4 space-x-2">
           <input type="hidden" name="format" value="turbo_stream"/>
