+++ app/views/projects/samples/shared/_metadata_toggle.html.erb
@@ -3,7 +3,7 @@
     :metadata,
     {
       checked: @search_params[:metadata].to_i == 1,
-      "data-action": "change->samples#toggleMetadata",
+      "data-action": "change->metadata-toggle#submit",
       class: "sr-only peer",
     },
   ) %>
